feat(frontend): consolidate credentials surfaces into /workspace/integrations/credentials (#227)
Merge api-keys, oauth-apps, and resource-tokens into a single page
using CategoryTabs (#225) with URL-sync via useTabParam (#230).

New files:
- credentials/page.tsx — thin orchestration (CategoryTabs + PageHeader)
- APIKeysTabPanel — self-contained API keys panel (from api-keys/page.tsx)
- OAuthAppsTabPanel — self-contained OAuth panel (from oauth-apps/page.tsx)
- ResourceTokensTabPanel — self-contained tokens panel (from resource-tokens/page.tsx)

Changes:
- Sidebar deep links: /integrations/{type} → /credentials?tab={type}
- Old routes (5 files) redirect to new consolidated route
- Cross-links updated (contexts settings page)
- i18n keys added (en + ja) with DX Lead recommended helpText
- helpText: purpose-first wording ("For programmatic access", not tool names)
