+++ frontend/src/app/(authenticated)/workspace/developer/api-keys/page.tsx
@@ -1,5 +1,5 @@
-import { redirect } from 'next/navigation';
+import { redirect } from "next/navigation";
 
 export default function ApiKeysRedirect() {
-  redirect('/workspace/integrations/api-keys');
+  redirect("/workspace/integrations/credentials?tab=api-keys");
 }
