+++ frontend/src/app/(authenticated)/workspace/developer/resource-tokens/page.tsx
@@ -1,5 +1,5 @@
-import { redirect } from 'next/navigation';
+import { redirect } from "next/navigation";
 
 export default function ResourceTokensRedirect() {
-  redirect('/workspace/integrations/resource-tokens');
+  redirect("/workspace/integrations/credentials?tab=resource-tokens");
 }
