+++ frontend/src/app/(authenticated)/workspace/oauth-apps/page.tsx
@@ -1,5 +1,5 @@
-import { redirect } from 'next/navigation';
+import { redirect } from "next/navigation";
 
 export default function OAuthAppsRedirect() {
-  redirect('/workspace/integrations/oauth-apps');
+  redirect("/workspace/integrations/credentials?tab=oauth-apps");
 }
